+++ playwright/lib/util.ts
@@ -164,7 +164,7 @@ const clearCookie = async (context: BrowserContext, cookieName: string) => {
 const fakeLogOut = async (context: BrowserContext) =>
 	await clearCookie(context, 'GU_U');
 
-const waitForSlot = async (page: Page, slot: string) => {
+const waitForSlot = async (page: Page, slot: string, waitForIframe = true) => {
 	const slotId = `#dfp-ad--${slot}`;
 	// create a locator for the slot
 	const slotLocator = page.locator(slotId);
@@ -172,10 +172,13 @@ const waitForSlot = async (page: Page, slot: string) => {
 	await slotLocator.isVisible();
 	// scroll to it
 	await slotLocator.scrollIntoViewIfNeeded();
-	// iframe locator
-	const iframe = page.locator(`${slotId} iframe`);
-	// wait for the iframe
-	await iframe.waitFor({ state: 'visible', timeout: 120000 });
+
+	if (waitForIframe) {
+		// iframe locator
+		const iframe = page.locator(`${slotId} iframe`);
+		// wait for the iframe
+		await iframe.waitFor({ state: 'visible', timeout: 120000 });
+	}
 };
 
 const waitForIsland = async (page: Page, island: string) => {
@@ -199,6 +202,38 @@ const countLiveblogInlineSlots = async (page: Page, isMobile: boolean) => {
 	return await page.locator(locator).count();
 };
 
+const getSlotName = (url: string) => {
+	const adRequest = new URL(url);
+	const adRequestParams = adRequest.searchParams;
+
+	const prevScp = new URLSearchParams(adRequestParams.get('prev_scp') ?? '');
+
+	return prevScp.get('slot') ?? 'unknown';
+};
+
+// Warn if any slots are unfilled
+const logUnfilledSlots = (page: Page) => {
+	page.on('response', (response) => {
+		const url = response.url();
+
+		const slotName = getSlotName(url);
+
+		if (url.includes('securepubads.g.doubleclick.net/gampad/ads')) {
+			const lineItemId = response.headers()['google-lineitem-id'] ?? '';
+			const creativeId = response.headers()['google-creative-id'] ?? '';
+
+			if (
+				!lineItemId ||
+				!creativeId ||
+				lineItemId === '-2' ||
+				creativeId === '-2'
+			) {
+				console.warn(`Unfilled slot: ${slotName}`);
+			}
+		}
+	});
+};
+
 export {
 	countLiveblogInlineSlots,
 	fakeLogOut,
@@ -207,4 +242,5 @@ export {
 	setupFakeLogin,
 	waitForIsland,
 	waitForSlot,
+	logUnfilledSlots,
 };