+++ playwright/tests/consent.spec.ts
@@ -1,16 +1,16 @@
 import type { Page } from '@playwright/test';
 import { expect, test } from '@playwright/test';
-import { articles } from '../../fixtures/pages';
-import type { GuPage } from '../../fixtures/pages/Page';
-import { cmpAcceptAll, cmpReconsent, cmpRejectAll } from '../../lib/cmp';
-import { loadPage } from '../../lib/load-page';
+import { articles } from '../fixtures/pages';
+import type { GuPage } from '../fixtures/pages/Page';
+import { cmpAcceptAll, cmpReconsent, cmpRejectAll } from '../lib/cmp';
+import { loadPage } from '../lib/load-page';
 import {
 	fakeLogOut,
 	getStage,
 	getTestUrl,
 	setupFakeLogin,
 	waitForSlot,
-} from '../../lib/util';
+} from '../lib/util';
 
 const { path } = articles[0] as unknown as GuPage;