+++ src/header-bidding/utils.ts
@@ -190,7 +190,7 @@ export const shouldIncludeCriteo = (): boolean => !isInAuOrNz();
 export const shouldIncludeSmart = (): boolean => isInUk() || isInRow();
 
 export const shouldIncludeKargo = (): boolean =>
-	isInUsa() && isInVariantSynchronous(prebidKargo, 'variant');
+	isInUsa() && !isInVariantSynchronous(prebidKargo, 'variant');
 
 export const shouldIncludeMobileSticky = once(
 	(): boolean =>
@@ -199,7 +199,7 @@ export const shouldIncludeMobileSticky = once(
 			min: 'mobile',
 			max: 'mobileLandscape',
 		}) &&
-			(isInUsOrCa() || isInAuOrNz()) &&
+			!isInUk() &&
 			window.guardian.config.page.contentType === 'Article' &&
 			!window.guardian.config.page.isHosted),
 );